Move sso provider type
