+++ workos/types/sso/sso_provider_type.py
@@ -1,0 +1,9 @@
+from typing import Literal
+
+
+SsoProviderType = Literal[
+    "AppleOAuth",
+    "GitHubOAuth",
+    "GoogleOAuth",
+    "MicrosoftOAuth",
+]
